requirements.txt
flask==2.0.1
psycopg2-binary==2.9.1
pillow==8.3.1
celery==5.1.2
redis==3.5.3
requests==2.26.0
pandas==1.3.2
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1